Move `show_script_editor`
