+++ client/ayon_aftereffects/api/lib.py
@@ -210,18 +210,3 @@ def publish_in_test(log, close_plugin_name=None):
                 close_plugin().process(context)
             except Exception as exp:
                 print(exp)
-
-
-def show_script_editor():
-    from ayon_core.tools.console_interpreter.ui import ConsoleInterpreterWindow
-
-    # Global so it doesn't get garbage collected instantly
-    global console_window
-    console_window = ConsoleInterpreterWindow()
-    console_window.setWindowTitle("Python Script Editor - AFX")
-    console_window.setWindowFlags(
-        console_window.windowFlags() |
-        QtCore.Qt.Dialog |
-        QtCore.Qt.WindowMinimizeButtonHint)
-    console_window.show()
-    console_window.raise_()
